+++ src/pages/OrderListPage/OrderListPage.tsx
@@ -1,25 +1,24 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { ThunkDispatch } from 'redux-thunk';
 import { Action } from 'redux';
 import { useSnackbar } from 'notistack';
 import PageHeader from 'components/shared/PageHeader/PageHeader';
 import PurchasedItem from 'components/units/PurchasedItem/PurchasedItem';
 import Spinner from 'components/shared/Spinner/Spinner';
-import * as T from 'types';
 import MESSAGE from 'constants/messages';
-import api from 'api';
-import { addCartItemRequest, getCartItemsRequest } from 'modules/cartItems/actions';
-import { CartState } from 'modules/cartItems/reducers';
 import { RootState } from 'modules';
+import { getCartItemsRequest } from 'modules/cartItems/actions';
+import useAddCartItem from 'hooks/useAddCartItem';
+import * as T from 'types';
+import api from 'api';
 import Styled from './OrderListPage.styles';
 
 const OrderListPage = () => {
-  const cartItems: CartState['cartItems'] = useSelector((state: RootState) => state.cartReducer.cartItems);
-
   const dispatch = useDispatch<ThunkDispatch<RootState, null, Action>>();
 
   const { enqueueSnackbar } = useSnackbar();
+  const addCartItem = useAddCartItem();
 
   const [isLoading, setLoading] = useState<boolean>(false);
   const [orders, setOrders] = useState<T.Order[]>([]);
@@ -38,22 +37,7 @@ const OrderListPage = () => {
   }, [enqueueSnackbar]);
 
   const handleClickCart = (product: T.Product) => {
-    if (isLoading || cartItems.status !== T.AsyncStatus.SUCCESS) return;
-
-    const cartItemIds = cartItems.data.map((cartItem) => cartItem.product.id);
-
-    if (cartItemIds.includes(product.id)) {
-      enqueueSnackbar(MESSAGE.EXIST_CART_ITEM);
-      return;
-    }
-
-    dispatch(addCartItemRequest(product))
-      .then(() => {
-        enqueueSnackbar(MESSAGE.ADDED_CART_ITEM_SUCCESS);
-      })
-      .catch((error: Error) => {
-        enqueueSnackbar(error.message);
-      });
+    addCartItem(product);
   };
 
   useEffect(() => {
