+++ src/pages/ProductDetailPage/ProductDetailPage.styles.ts
@@ -61,6 +61,7 @@ const CartButton = styled.button`
   text-align: center;
   border: none;
   cursor: pointer;
+  outline: none;
 `;
 
 export default {
